Refactor error handling for command execution
Apply Upstream patch from Michael Meyer <mijomeyer@t-online.de>
which consolidates further methods mkdir and chmod in the
common space and also moves error handling out of the
podman pilot into the common space
