+++ common/src/user.rs
@@ -1,7 +1,29 @@
+//
+// Copyright (c) 2023 Elektrobit Automotive GmbH
+//
+// This file is part of flake-pilot
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
 use std::{process::Command, ffi::OsStr};
-
 use serde::{Serialize, Deserialize};
-
+use crate::command::{CommandExtTrait, CommandError};
 
 #[derive(Debug, Default, Clone, Copy, Serialize, Deserialize)]
 pub struct User<'a> {
@@ -25,4 +47,20 @@ impl<'a> User<'a> {
         c.arg(command);
         c
     }
+}
+
+pub fn chmod(filename: &str, mode: &str, user: User) -> Result<(), CommandError> {
+    /*!
+    Chmod filename via sudo
+    !*/
+    user.run("chmod").arg(mode).arg(filename).perform()?;
+    Ok(())
+}
+
+pub fn mkdir(dirname: &str, mode: &str, user: User) -> Result<(), CommandError> {
+    /*!
+    Make directory via sudo
+    !*/
+    user.run("mkdir").arg("-p").arg("-m").arg(mode).arg(dirname).perform()?;
+    Ok(())
 }
